Added the ability to add folders.
Separated the bloc structure and lowered the total file count
