+++ wordify/lib/features/word_tree/domain/entities/dictionary.dart
@@ -3,7 +3,7 @@ import 'package:wordify/core/util/stack_linked_hash_map.dart';
 import 'package:wordify/features/word_tree/domain/entities/folder.dart';
 
 ///A dictionary of the app. It contains a:
-///list of folders that are in view (could be opened)
+///tree folders that are in view (could be/are opened)
 ///list of active folders (the folders that have been opened and are in view)
 ///set of all folders which have ever been opened (a caching mechanism)
 class Dictionary {
