+++ wordify/lib/features/word_tree/domain/entities/word.dart
@@ -5,7 +5,7 @@ class Word {
 
 
   const Word({
-    this.word = '',
-    this.translation = ''
+    required this.word,
+    required this.translation
   });
 }
