+++ wordify/lib/features/word_tree/domain/repositories/folder_repository.dart
@@ -3,7 +3,7 @@ import 'package:wordify/features/word_tree/domain/entities/folder.dart';
 ///A source-ambiguous storage access point interface.
 abstract class FolderRepository {
   ///Create a new folder.
-  Future<Folder> addFolder(Folder folder);
+  Future<Folder> addFolder(Folder? parentFolder, Folder folder);
   ///Update a folder.
   Future<Folder> updateFolder(Folder oldFolder, Folder newFolder);
   ///Delete a folder.
