+++ wordify/lib/features/word_tree/presentation/pages/main_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:wordify/core/ui_kit/buttons.dart';
-import 'package:wordify/features/word_tree/presentation/pages/create_word_template_screen.dart';
-import 'package:wordify/features/word_tree/presentation/widgets/folder_list_widget.dart';
-import 'package:wordify/features/word_tree/presentation/widgets/word_list_widget.dart';
+import 'package:wordify/features/word_tree/presentation/pages/word_template_screen.dart';
+import 'package:wordify/features/word_tree/presentation/widgets/folder_view_widget.dart';
+import 'package:wordify/features/word_tree/presentation/widgets/word_view_widget.dart';
 
 
 ///
@@ -17,11 +17,11 @@ class MainScreen extends StatelessWidget {
         children: [
           Expanded( //Folders column
             flex: 25,
-            child: FolderListWidget()
+            child: FolderViewWidget()
           ),
           Expanded( //Words column
             flex: 75,
-            child: WordListWidget()
+            child: WordViewWidget()
           )
         ]
       ),
