+++ wordify/lib/features/word_tree/presentation/state_management/validation_bloc.dart
@@ -1,5 +1,6 @@
 import 'package:wordify/features/word_tree/domain/entities/folder.dart';
 import 'package:wordify/features/word_tree/domain/validation/folder_validation.dart';
+import 'package:wordify/features/word_tree/domain/validation/word_validation.dart';
 
 ///Validate the update of a folder.
 class FolderValidationBloc {
@@ -26,5 +27,33 @@ class FolderValidationBloc {
 
   String? validateChooseFolder(Folder? folder) {
     return _folderValidator.validateChooseFolder(folder);
+  }
+}
+
+
+
+///Validate the update of a word.
+class WordValidationBloc {
+  static final WordValidationBloc _instance = WordValidationBloc._internal();
+  final WordValidation _wordValidator = WordValidation();
+
+
+  factory WordValidationBloc() {
+    return _instance;
+  }
+
+
+  WordValidationBloc._internal();
+
+
+  ///
+  String? validateWordWord(String word) {
+    return _wordValidator.validateWord(word);
+  }
+
+
+  ///
+  String? validateWordTranslation(String translation) {
+    return _wordValidator.validateTranslation(translation);
   }
 }
